backend/requirements.txt
asgiref==3.8.1
Django==5.0.3
django-cors-headers==4.3.1
django-debug-toolbar==4.3.0
django-filter==24.2
djangorestframework==3.15.1
psycopg2-binary==2.9.9
setuptools==69.1.1
sqlparse==0.4.4
tzdata==2024.1
wheel==0.43.0
